Add player attacking and add unique colors to entities

+++ Server/src/GameObjects/Wall.cpp
@@ -2,14 +2,14 @@
 
 #include "OBJLoader.h"
 
-Wall::Wall(GroundPlane &gp, const int x, const int y, const glm::vec3 scale)
+Wall::Wall(GroundPlane &gp, const double x, const double y, const glm::vec3 scale)
 	: Entity("unitcube", false, false)
 {
 	DoScale(scale);
 
-	float xPos = -gp.GetWidth() / 2 + x + scale.x / 2;
-	float yPos = gp.GetHeight() / 2 - y - scale.y / 2;
-	float zPos = scale.z / 2;
+	double xPos = -gp.GetWidth() / 2 + x + scale.x / 2;
+	double yPos = gp.GetHeight() / 2 - y - scale.y / 2;
+	double zPos = scale.z / 2;
 
 	SetPosition(glm::vec3(xPos, yPos, zPos));
 }
