+++ Server/src/GameObjects/Wall.h
@@ -7,7 +7,7 @@
 class Wall : public Entity
 {
 public:
-	Wall(GroundPlane &gp, const int x, const int y, const glm::vec3 scale);
+	Wall(GroundPlane &gp, const double x, const double y, const glm::vec3 scale);
 	~Wall();
 
 	const std::string &ToProtocolString() override;
